Added tests for send, staking
